ImplicitDomainQuadrature.jl

ImplicitDomainQuadrature.AbstractBasis

ImplicitDomainQuadrature.AbstractBasis1D

ImplicitDomainQuadrature.InterpolatingPolynomial

ImplicitDomainQuadrature.InterpolatingPolynomial

ImplicitDomainQuadrature.InterpolatingPolynomial

ImplicitDomainQuadrature.LagrangePolynomialBasis

ImplicitDomainQuadrature.LagrangePolynomialBasis

ImplicitDomainQuadrature.QuadratureRule

ImplicitDomainQuadrature.ReferenceQuadratureRule

ImplicitDomainQuadrature.SignSearch

ImplicitDomainQuadrature.TensorProductBasis

ImplicitDomainQuadrature.TensorProductBasis

Base.:==

Base.:==

Base.getindex

Base.getindex

Base.isequal

Base.isequal

Base.iterate

Base.iterate

Base.muladd

Base.sign

Base.sign

Base.sign

BranchAndPrune.bisect

BranchAndPrune.process

ImplicitDomainQuadrature.coefficient_number_type

ImplicitDomainQuadrature.derivative

ImplicitDomainQuadrature.extend

ImplicitDomainQuadrature.gradient

ImplicitDomainQuadrature.gradient

ImplicitDomainQuadrature.gradient

ImplicitDomainQuadrature.interpolation_points

ImplicitDomainQuadrature.isSuitable

ImplicitDomainQuadrature.lagrange_polynomial

ImplicitDomainQuadrature.lagrange_polynomials

ImplicitDomainQuadrature.min_diam

ImplicitDomainQuadrature.normalization

ImplicitDomainQuadrature.polynomial_from_roots

ImplicitDomainQuadrature.quadrature

ImplicitDomainQuadrature.quadrature_transformers

ImplicitDomainQuadrature.roots_and_ends

ImplicitDomainQuadrature.run_search

ImplicitDomainQuadrature.signConditionsSatisfied

ImplicitDomainQuadrature.surface_quadrature

ImplicitDomainQuadrature.tensorProduct

ImplicitDomainQuadrature.tensorProductPoints

ImplicitDomainQuadrature.transform

ImplicitDomainQuadrature.unique_root_intervals

ImplicitDomainQuadrature.unique_roots

ImplicitDomainQuadrature.update!

ImplicitDomainQuadrature.update!

ImplicitDomainQuadrature.value_and_derivative

ImplicitDomainQuadrature.InterpolatingPolynomial — Type

InterpolatingPolynomial{N,NFuncs,B<:AbstractBasis,T}

interpolate a VECTOR of N with a basis B composed of NFuncs functions

Fields

- `coeffs::SMatrix{N,NFuncs,T}`
- `basis::B`

ImplicitDomainQuadrature.InterpolatingPolynomial — Method

(P::InterpolatingPolynomial{1})(x...)

evaluate P at x, the result is a scalar (P::InterpolatingPolynomial)(x...) evaluate P at x (P::InterpolatingPolynomial)(x::AbstractVector) evaluate P at the point vector x

ImplicitDomainQuadrature.InterpolatingPolynomial — Method

InterpolatingPolynomial(T::Type{<:Number}, N::Int,
    basis::AbstractBasis{NFuncs}) where {NFuncs}

initialize an InterpolatingPolynomial{N,NFuncs,T} object with coefficients zeros(T,N,NFuncs) InterpolatingPolynomial(N::Int, basis::AbstractBasis) initialize an InterpolatingPolynomial object with Float64 coefficients. InterpolatingPolynomial(N::Int, dim::Int, order::Int, start::T, stop::T) where {T<:Real} initialize a dim dimensional basis of order order and pass this to the InterpolatingPolynomial constructor. InterpolatingPolynomial(N::Int, dim::Int, order::Int; start = -1.0, stop = 1.0) initialize a dim dimensional basis of order order and pass this to the InterpolatingPolynomial constructor.

ImplicitDomainQuadrature.QuadratureRule — Type

QuadratureRule{D,T}

a quadrature rule in D spatial dimensions. This is a mutable type; points and weights can be added via the update! method.

Inner Constructor:

QuadratureRule(points::Matrix{T}, weights::Vector{T}) where {T<:Real}

Outer Constructors:

QuadratureRule(T::Type{<:Real}, dim::Int)

initializes (dim,0) matrix of points and (0) vector of weights. Both are of type T QuadratureRule(dim::Int) initializes (dim,0) matrix of points and (0) vector of weights of type Float64

ImplicitDomainQuadrature.TensorProductBasis — Type

TensorProductBasis{N,NFuncs,T<:AbstractBasis1D} <: AbstractBasis{NFuncs}

construct an N dimensional polynomial basis by tensor product of T with a total of NFuncs functions. Note that NFuncs can be inferred from T{nfuncs} as NFuncs = nfuncs^N.

Fields

- `basis::T` the underlying 1D basis
- `points::SMatrix{N,NFuncs}` a matrix of point vectors

ImplicitDomainQuadrature.TensorProductBasis — Method

TensorProductBasis(dim::Int, order::Int; start = -1.0, stop = 1.0)

construct a dim dimensional polynomial basis with variable x using a tensor product of order order LagrangePolynomialBasis. The polynomials are equispaced from start to stop.

Base.:== — Method

==(b1::LagrangePolynomialBasis{NF},b2::LagrangePolynomialBasis{NF}) where {NF}

returns true if b1 and b2 have the same funcs and points.

Base.:== — Method

==(tp1::TensorProductBasis{D,T,NF}, tp2::TensorProductBasis{D,T,NF}) where {D,T,NF}

returns true if tp1 and tp2 have identical basis and points, false otherwise.

Base.isequal — Method

Base.isequal(b1::LagrangePolynomialBasis{NF},b2::LagrangePolynomialBasis{NF}) where {NF}

returns true if b1 and b2 have the same funcs and points.

Base.isequal — Method

Base.isequal(tp1::TensorProductBasis{D,T,NF}, tp2::TensorProductBasis{D,T,NF}) where {D,T,NF}

returns true if tp1 and tp2 have identical basis and points, false otherwise.

Base.sign — Method

Base.sign(m::Int,s::Int,S::Bool,sigma::Int)

See sec. 3.2.4 of Robert Saye's 2015 SIAM paper

Base.sign — Method

Base.sign(P::InterpolatingPolynomial{1}, int::IntervalBox; algorithm = :TaylorModels, tol = 1e-3, order = 5)

special function for an interpolating polynomial type.

Base.sign — Method

sign(f, int::Interval)

return

+1 if f is uniformly positive on int

-1 if f is uniformly negative on int

0 if f has at least one zero crossing in int (f assumed continuous)

ImplicitDomainQuadrature.gradient — Method

gradient(P::InterpolatingPolynomial)(x)

return the gradient of P evaluated at x. gradient(P::InterpolatingPolynomial, dir::Int, x...) return the gradient of P along direction dir evaluated at x. gradient(P::InterpolatingPolynomial, x::AbstractVector) evaluate the gradient of P at the point vector x

ImplicitDomainQuadrature.gradient — Method

gradient(B::TensorProductBasis{N}, dir::Int, x::AbstractVector) where {N}

evaluate the gradient of B along direction dir at the point vector x gradient(B::TensorProductBasis{dim}, x::AbstractVector) where {dim} return an (N,dim) matrix such that the Ith row is the gradient of basis I. Here N is the total number of basis functions. gradient(B::TensorProductBasis{3}, dir::Int, x::Number, y::Number, z::Number) evaluate the gradient of B at (x,y,z) along direction dir. gradient(B::TensorProductBasis{3}, x::Number, y::Number, z::Number) returns an (N,3) matrix, where row I is the (2D) gradient vector of the Ith basis function. Here N is the total number of basis functions. gradient(B::TensorProductBasis{2}, x::Number, y::Number) returns an (N,2) matrix, where row I is the (2D) gradient vector of the Ith basis function. Here N is the total number of basis functions.

ImplicitDomainQuadrature.gradient — Method

gradient(B::LagrangePolynomialBasis, x::Number)

alias for derivative(B, x)

ImplicitDomainQuadrature.interpolation_points — Method

interpolation_points(B::TensorProductBasis{N,NFuncs,T}) where {NFuncs,T}

return the interpolation points of the tensor product basis

ImplicitDomainQuadrature.quadrature — Method

quadrature(F::Vector, sign_conditions::Vector{Int}, lo::T, hi::T, quad1d::ReferenceQuadratureRule{N,T}) where {N,T}

return a 1D quadrature rule that can be used to integrate in the domain where each f in F has sign specified in sign_conditions in the interval [lo,hi] quadrature(F::Vector, signconditions::Vector{Int}, int::Interval{T}, quad1d::ReferenceQuadratureRule{N,T}) where {N,T} conveniance function for when the interval over which the quadrature rule is required is specified by an Interval type quadrature(F::Vector, signconditions::Vector{Int}, heightdir::Int, lo::T, hi::T, x0::AbstractVector{T}, w0::T, quad1d::ReferenceQuadratureRule{N,T}) where {N,T} return a tuple points,weights representing quadrature points and weights obtained by transforming quad1d into an interval bounded by x0 and the zero level set of F by extending along `heightdir` quadrature(F::Vector, signconditions::Vector{Int}, heightdir::Int, lo::T, hi::T, x0::AbstractMatrix, w0::AbstractVector, quad1d::ReferenceQuadratureRule{N,T}) where {N,T}

ImplicitDomainQuadrature.update! — Method

update!(P::InterpolatingPolynomial, coeffs::AbstractMatrix)

update P.coeffs = coeffs

ImplicitDomainQuadrature.update! — Method

update!(quad::QuadratureRule{D,T}, points::AbstractMatrix{T}, weights::AbstractVector{T}) where {D,T}

concatenates points and weights into quad.points and quad.weights and increments quad.N appropriately. update!(quad::QuadratureRule{D,T}, point::AbstractVector{T}, weight::T) where {D,T} adds a single point,weight pair to quad

ImplicitDomainQuadrature.AbstractBasis — Type

AbstractBasis{N}

abstract supertype for a function basis with N functions

ImplicitDomainQuadrature.AbstractBasis1D — Type

AbstractBasis1D{N}

abstract supertype for a 1D function basis with N functions

ImplicitDomainQuadrature.LagrangePolynomialBasis — Type

LagrangePolynomialBasis{N} <: AbstractBasis1D{N}

A basis of N Lagrange polynomials. The polynomial order is N - 1.

Fields

- `funcs::PolynomialSystem{N,1}` - a system of static polynomials
- `points::StaticVector{N}` - a static vector of support points

ImplicitDomainQuadrature.LagrangePolynomialBasis — Method

LagrangePolynomialBasis(order::Int, start::T, stop::T) where {T<:Real}

construct a polynomial basis with variable x of order order with equally spaced points between start and stop. LagrangePolynomialBasis(order::Int) construct a polynomial basis with variable x of order order with equally spaced points of type T between -1.0 and 1.0

ImplicitDomainQuadrature.ReferenceQuadratureRule — Type

ReferenceQuadratureRule{N,T}

a one-dimensional quadrature rule defined on the reference domain [-1,1].

Inner Constructors:

ReferenceQuadratureRule(points::SMatrix{1,N,T}, weights::SVector{N,T}) where {T<:Real}

Outer Constructors:

ReferenceQuadratureRule(points::AbstractMatrix, weights::AbstractVector)
ReferenceQuadratureRule(points::AbstractVector, weights::AbstractVector)
ReferenceQuadratureRule(N::Int)

return an N point gauss legendre quadrature rule on [-1,1]

ImplicitDomainQuadrature.SignSearch — Type

SignSearch{N,T} <: AbstractBreadthFirstSearch{IntervalBox{N,T}}

a BranchAndPrune search type that is used to determine the sign of a function over a given interval

Base.getindex — Method

Base.getindex(quad::QuadratureRule{D}, i::Int) where {D}

return (p,w) where p is an appropriate view into the ith quadrature point, and w is the ith quadrature weight.

Base.getindex — Method

Base.getindex(quad::ReferenceQuadratureRule{N}, i::Int) where {N}

return the ith point weight pair p,w.

Base.iterate — Method

Base.iterate(quad::ReferenceQuadratureRule{N}, state=1)

return each point, weight pair as a tuple (p,w)

Base.iterate — Method

Base.iterate(quad::QuadratureRule{D}, state=1) where {D}

return each point, weight pair as a tuple (p,w) where p is an appropriate view into the point matrix quad.points

Base.muladd — Method

Base.muladd(tm::TaylorModelN, a::Number, b::Number)

overloading muladd to avoid an error

BranchAndPrune.bisect — Method

BranchAndPrune.bisect(::SignSearch, interval::IntervalBox)

split the given interval along its largest dimension

BranchAndPrune.process — Method

BranchAndPrune.process(search::SignSearch, interval::IntervalBox)

process the given interval to determine the sign of search.f in this interval

ImplicitDomainQuadrature.coefficient_number_type — Method

coefficient_number_type(B::LagrangePolynomialBasis)
coefficient_number_type(B::TensorProductBasis)

return the type of the coefficient, i.e. Float64, Float32

ImplicitDomainQuadrature.derivative — Method

derivative(B::LagrangePolynomialBasis, x::Number)

evaluate the derivative of basis B at the point x

ImplicitDomainQuadrature.extend — Method

extend(x0::Number, k::Int, x::Number)

extend x0 into 2D space by inserting x in direction k extend(x0::AbstractVector, k::Int, x::Number) extend the point vector x0 into d+1 dimensional space by using x as the kth coordinate value in the new point vector extend(x0::AbstractVector, k::Int, x::AbstractMatrix) extend the point vector x0 along direction k treating x as the new coordinate values

ImplicitDomainQuadrature.isSuitable — Method

isSuitable(height_dir::Int, F::AbstractVector, box::IntervalBox)

checks if the given height_dir is a suitable height direction for each f in F in the domain specified by box

ImplicitDomainQuadrature.lagrange_polynomial — Method

lagrange_polynomial(x::DP.PolyVar,roots::AbstractVector{T},index::Int) where {T}

return the indexth lagrange polynomial from the vector of roots with variable x.

The polynomial is normalized such that it evaluates to 1.0 on roots[index].

Examples

julia> import DynamicPolynomials.@polyvar
julia> @polyvar x
(x,)
julia> ImplicitDomainQuadrature.lagrange_polynomial(x,[-1,0,1],2)
-x² + 1.0

See also: lagrange_polynomials

ImplicitDomainQuadrature.lagrange_polynomials — Method

lagrange_polynomials(x::DP.PolyVar,roots::AbstractVector{T}) where {T}

return a vector of lagrange polynomial bases constructed from the given roots.

Examples

julia> import DynamicPolynomials.@polyvar
julia> @polyvar x
(x,)
julia> ImplicitDomainQuadrature.lagrange_polynomials(x, [-1,0,1])
3-element Array{DynamicPolynomials.Polynomial{true,Float64},1}:
 0.5x² - 0.5x
 -x² + 1.0
 0.5x² + 0.5x

See also: lagrange_polynomial

ImplicitDomainQuadrature.min_diam — Method

min_diam(box::IntervalBox)

return the minimum diameter of box

ImplicitDomainQuadrature.normalization — Method

normalization(roots::AbstractVector{T},skipindex::Int) where {T}

return the denominator of the standard Lagrange polynomial formula.

Dividing f = polynomial_from_roots(x,roots,skipindex)/normalization(roots,skipindex) gives a normalized polynomial which evaluates to 1.0 on the skipindexth root i.e. f(roots[skipindex] = 1.0.

Examples

julia> ImplicitDomainQuadrature.normalization([-1,1],1)
-2

ImplicitDomainQuadrature.polynomial_from_roots — Method

polynomial_from_roots(x::DP.PolyVar,roots::AbstractVector{T},skipindex::Int) where {T}

return a DynamicPolynomial with variable x with zeros at roots but skipping the skipindexth root.

Examples

julia> import DynamicPolynomials.@polyvar
julia> @polyvar x
(x,)
julia> ImplicitDomainQuadrature.polynomial_from_roots(x,[-1,1],1)
x - 1

ImplicitDomainQuadrature.quadrature_transformers — Method

quadrature_transformers(lo, hi)

return 0.5(hi - lo), 0.5*(hi + lo) which are the factors used to transform a quadrature rule from (-1,1) to (lo,hi)

ImplicitDomainQuadrature.roots_and_ends — Method

roots_and_ends(F::Vector, x1::T, x2::T)

return a sorted array containing [x1, <roots of each f in F>, x2]

ImplicitDomainQuadrature.run_search — Method

run_search(f, interval, algorithm, tol, order)

run a BranchAndPrune search until the sign of f in the given interval is determined

ImplicitDomainQuadrature.signConditionsSatisfied — Method

signConditionsSatisfied(funcs,xc,sign_conditions)

return true if each f in funcs evaluated at xc has the sign specified in sign_conditions. Note: if sign_conditions[i] == 0 then the function always returns true.

ImplicitDomainQuadrature.surface_quadrature — Method

surface_quadrature(F::InterpolatingPolynomial, height_dir::Int, lo::T, hi::T, x0::AbstractVector{T}, w0::T) where {T}

return a point,weight pair representing a quadrature point and weight that is obtained by projecting x0 along height_dir onto the zero level set of F within the interval (lo,hi). weight is scaled appropriately by the curvature of F surfacequadrature(F::InterpolatingPolynomial, heightdir::Int, int::Interval{T}, x0::AbstractVector{T}, w0::T) where {T} convenience function when the interval within which to project x0 is defined by an Interval type surfacequadrature(F::InterpolatingPolynomial, heightdir::Int, lo::T, hi::T, x0::AbstractMatrix{T}, w0::AbstractVector{T}) where {T} project each column of x0 onto the zero level set of F surfacequadrature(F::InterpolatingPolynomial, heightdir::Int, int::Interval{T}, x0::AbstractMatrix{T}, w0::AbstractVector{T}) where {T} quadrature(P::InterpolatingPolynomial{N,NF,B,T}, sign_condition::Int, surface::Bool, box::IntervalBox{2,T}, quad1d::ReferenceQuadratureRule{NQ,T}) where {B<:TensorProductBasis{2}} where {N,NF,NQ,T} corresponds to Algorithm 3 of Robert Saye's 2015 SIAM paper

ImplicitDomainQuadrature.tensorProduct — Method

tensorProduct(quad1d::ReferenceQuadratureRule, box::IntervalBox{2})

returns a quadrature rule representing the "tensor product" or "kroneker product" of quad1d that is appropriately transformed into box

ImplicitDomainQuadrature.tensorProductPoints — Method

tensorProductPoints(p1::Matrix, p2::Matrix)

returns a matrix of points representing the "tensor product" (or kroneker product) of p1 and p2

ImplicitDomainQuadrature.transform — Method

transform(quad::ReferenceQuadratureRule{N,T}, lo::T, hi::T) where {N,T}

transform quad from the interval [-1,1] to the interval [a,b] transform(quad::ReferenceQuadratureRule, int::Interval) transform quad from the interval [-1,1] to int

ImplicitDomainQuadrature.unique_root_intervals — Method

unique_root_intervals(f, x1::T, x2::T) where {T<:Real}

returns sub-intervals in [x1,x2] which contain unique roots of f

ImplicitDomainQuadrature.unique_roots — Method

unique_roots(f, x1::T, x2::T) where {T<:Real}

returns the unique roots of f in the interval (x1, x2)

ImplicitDomainQuadrature.value_and_derivative — Method

value_and_derivative(B::LagrangePolynomialBasis, x::Number)

return (v,d) where v is a vector of the polynomial basis values and d is a vector of the derivative values of the polynomial basis at the point x.